Redo project folder structure to be more organized.
Also changed a few items on the cook list for better testing and updated the build file.

Cleaned up some warnings and unnecessary syntax.
