+++ app/build.gradle
@@ -41,7 +41,7 @@ dependencies {
     implementation libs.retrofit
     implementation libs.converter.gson
     implementation libs.glide
-    implementation 'com.squareup.okhttp3:logging-interceptor:4.9.1'
+    implementation libs.logging.interceptor
     testImplementation libs.junit
     androidTestImplementation libs.ext.junit
     androidTestImplementation libs.espresso.core
