+++ app/src/main/AndroidManifest.xml
@@ -15,22 +15,22 @@
         android:theme="@style/Theme.KitchInventory"
         tools:targetApi="31">
         <activity
-            android:name=".RecipesActivity"
+            android:name=".activities.RecipesActivity"
             android:exported="false" />
         <activity
-            android:name=".CookActivity"
+            android:name=".activities.CookActivity"
             android:exported="false" />
         <activity
-            android:name=".KitchenActivity"
+            android:name=".activities.KitchenActivity"
             android:exported="false" />
         <activity
-            android:name=".ShoppingListActivity"
+            android:name=".activities.ShoppingListActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".activities.LoginActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
