+++ app/src/main/java/com/beaconcode/kitchinventory/activities/MainActivity.java
@@ -1,4 +1,4 @@
-package com.beaconcode.kitchinventory;
+package com.beaconcode.kitchinventory.activities;
 
 import android.content.Context;
 import android.content.Intent;
@@ -6,13 +6,9 @@ import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuInflater;
 
-import androidx.activity.EdgeToEdge;
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
 
+import com.beaconcode.kitchinventory.R;
 import com.beaconcode.kitchinventory.databinding.ActivityMainBinding;
 
 /**
