+++ app/src/main/java/com/beaconcode/kitchinventory/activities/ShoppingListActivity.java
@@ -1,4 +1,4 @@
-package com.beaconcode.kitchinventory;
+package com.beaconcode.kitchinventory.activities;
 
 import android.content.Context;
 import android.content.Intent;
@@ -7,6 +7,8 @@ import android.view.Menu;
 import android.view.MenuInflater;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.beaconcode.kitchinventory.R;
 
 /**
  * This activity will display the current items in the user's shopping list from the database.
@@ -18,7 +20,7 @@ public class ShoppingListActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_shopping_list);
-        };
+        }
 
     /**
      * Initialize the contents of the Activity's standard options menu.
