+++ app/src/main/java/com/beaconcode/kitchinventory/data/database/KitchenDAO.java
@@ -1,12 +1,11 @@
-package com.beaconcode.kitchinventory.database;
+package com.beaconcode.kitchinventory.data.database;
 
 import androidx.room.Dao;
 import androidx.room.Delete;
 import androidx.room.Insert;
-import androidx.room.Query;
 import androidx.room.Update;
 
-import com.beaconcode.kitchinventory.database.entities.Kitchen;
+import com.beaconcode.kitchinventory.data.database.entities.Kitchen;
 
 /**
  * Kitchen DAO
