+++ app/src/main/java/com/beaconcode/kitchinventory/data/database/entities/ShoppingList.java
@@ -1,19 +1,18 @@
-package com.beaconcode.kitchinventory.database.entities;
+package com.beaconcode.kitchinventory.data.database.entities;
 
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
 
-import com.beaconcode.kitchinventory.database.KitchenDatabase;
+import com.beaconcode.kitchinventory.data.database.KitchenDatabase;
 
 import java.util.Objects;
 
 /**
- * Kitchen Class
- * Organizes the data of ingredients that user currently has in kitchen
+ * POJO representing a shopping list
  */
 
-@Entity(tableName = KitchenDatabase.KITCHEN_TABLE)
-public class Kitchen {
+@Entity(tableName = KitchenDatabase.SHOPPING_LIST_TABLE)
+public class ShoppingList {
 
     @PrimaryKey(autoGenerate = true)
     private int inventoryId;
@@ -24,16 +23,25 @@ public class Kitchen {
     private int quantity;
     private double price;
 
-
-    //constructor
-    public Kitchen(String name, int quantity, double price) {
+    public ShoppingList(String name, int quantity, double price) {
         this.name = name;
         this.quantity = quantity;
         this.price = price;
     }
 
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        ShoppingList shoppingList = (ShoppingList) o;
+        return inventoryId == shoppingList.inventoryId && foodId == shoppingList.foodId && userId == shoppingList.userId;
+    }
 
-    //getters and setters
+    @Override
+    public int hashCode() {
+        return Objects.hash(inventoryId, foodId, userId);
+    }
+
     public int getInventoryId() {
         return inventoryId;
     }
@@ -56,20 +64,6 @@ public class Kitchen {
 
     public void setUserId(int userId) {
         this.userId = userId;
-    }
-
-    //equals and hashcode
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Kitchen kitchen = (Kitchen) o;
-        return inventoryId == kitchen.inventoryId && foodId == kitchen.foodId && userId == kitchen.userId;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(inventoryId, foodId, userId);
     }
 
     public String getName() {
@@ -96,3 +90,5 @@ public class Kitchen {
         this.price = price;
     }
 }
+
+
