+++ app/src/main/java/com/beaconcode/kitchinventory/data/model/Meal.java
@@ -1,4 +1,4 @@
-package com.beaconcode.kitchinventory;
+package com.beaconcode.kitchinventory.data.model;
 
 import java.util.Objects;
