+++ app/src/main/res/layout/activity_cook.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CookActivity">
+    tools:context=".activities.CookActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_cook"
