+++ app/src/main/res/layout/activity_recipes.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RecipesActivity">
+    tools:context=".activities.RecipesActivity">
 
     <TextView
         android:id="@+id/tv_ingredientName"
